requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
dj-database-url==0.5.0
Django==4.1.7
django-cors-headers==3.13.0
django-soft-delete==0.9.21
djangorestframework==3.14.0
idna==3.4
jsonfield==3.1.0
methodism==0.0.1
python-dotenv==1.0.0
pytz==2023.2
requests==2.28.2
sqlparse==0.4.3
tzdata==2023.2
urllib3==1.26.15
whitenoise==6.4.0
